Player can now take multiple actions on the action gauge
Currently random enemies are damaged but that behaviour will eventually change.

+++ Assets/Scripts/Items/Weapon.cs
@@ -8,8 +8,8 @@ public class Weapon : Item
     public float atp;
     public float mag;
     public ActionGaugeData actGauge;
-    short defaultTokenCount {get;} = 3;
-    public short tokenCount = 3;    //default amount is 3
+    int defaultTokenCount {get;} = 3;
+    public int tokenCount = 3;    //default amount is 3
     public Skill weaponSkill;
 
     public enum WeaponType
